All types of groups show up in search group facet
With the addition of open and restricted groups the group facet
needs to be modified to accomodate these group types. Previous
behavior was only private groups show up in the search auto-
complete. Now any group associated with the user will show up.
This means any group where the user is a member or group creator.
These are attained by calling associated_groups of the
list_group service.